pyproject.toml
[tool.poetry]
name = "semadec-hub"
version = "0.1.0"
description = "Projeto desenvolvido com a finalidade de facilitar o gerenciamento da SEMADEC no IFRN Campus Currais Novos."
authors = ["Joamerson Islan <joamersonislan38@gmail.com>, Sérgio Dantas <sergiodnts828@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
Django = "~5.0.1"
python-dotenv = "~1.0.0"
Pillow = "~10.2.0"
social-auth-app-django = "~5.4.0"
django-environ = "~0.11.2"
django-summernote = "~0.8.20.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.4"
pytest-django = ">=4.8.0"
coverage = ">=7.4.0"
ruff = ">=0.1.14"
factory-boy = ">=3.3.0"
faker = ">=22.7.0"
django-debug-toolbar = ">=4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = [
    "test.py", "test_*.py", "*_test.py",
    "tests.py", "tests_*.py", "*_tests.py",
    "testing/python/*.py"
    ]
addopts = ["--doctest-modules", "--strict-markers", "-rP"]
markers = ["slow", "fast"]
filterwarnings = [
    "error",
    "ignore::pytest.PytestCollectionWarning",
]

[tool.ruff]
line-length = 88
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["F401"]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.coverage.run]
branch = true
omit = [
    "*/venv/*",
    "*/*test*.py",
    "*/conftest.py",
    "*/factory.py",
    "manage.py",
    "core/*.py",
    "utils/*.py",
    "**/__init__.py",
    "*/migrations/*.py",
]

[tool.coverage.html]
directory = "coverage"
